fix: some changes after review, update axios package version
