+++ packages/extensions/package.json
@@ -30,7 +30,7 @@
   },
   "dependencies": {
     "@nestjs/common": "^10.0.0",
-    "axios": "^0.27.2",
+    "axios": "^0.28.1",
     "@playwright/test": "^1.51.1",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.5.6",
