+++ packages/pg-reporter/package.json
@@ -30,6 +30,6 @@
   },
   "dependencies": {
     "@playwright/test": "^1.51.1",
-    "axios": "^0.27.2"
+    "axios": "^0.28.1"
   }
 }
