+++ wallets-testing/pages/widget.page.ts
@@ -4,6 +4,7 @@ export class WidgetPage {
   connectBtn: Locator;
   stakeInput: Locator;
   stakeSubmitBtn: Locator;
+  enabledStakeSubmitBtn: Locator;
 
   headerAccountSection: Locator;
   providerName: Locator;
@@ -15,6 +16,9 @@ export class WidgetPage {
     this.connectBtn = this.page.getByTestId('connectBtn');
     this.stakeInput = this.page.getByTestId('stakeInput');
     this.stakeSubmitBtn = this.page.getByTestId('stakeSubmitBtn');
+    this.enabledStakeSubmitBtn = this.page.locator(
+      'button[data-testid="stakeSubmitBtn"]:not([disabled])',
+    );
     this.headerAccountSection = this.page.getByTestId('accountSectionHeader');
     this.providerName = this.page.locator('div[data-testid="providerName"]');
     this.ethAvailableToStakeValue = this.page.getByTestId(
@@ -24,9 +28,21 @@ export class WidgetPage {
     this.copyWcUrlBtn = this.page.locator('.wcm-action-btn');
   }
 
+  async goto(path?: string) {
+    await this.page.goto(path);
+  }
+
   async getWalletButtonByName(walletButtonName: string) {
     return this.page.getByRole('button').getByText(walletButtonName, {
       exact: true,
     });
+  }
+
+  async waitForPage(timeout?: number) {
+    return this.page.context().waitForEvent('page', { timeout: timeout });
+  }
+
+  async closeAccountModal() {
+    await this.page.locator('div[role="dialog"] button').nth(0).click();
   }
 }
