+++ wallets-testing/services/index.ts
@@ -1,2 +1,1 @@
 export * from './service.widget';
-export * from './service.test';
