+++ wallets-testing/services/service.widget.ts
@@ -1,28 +1,26 @@
 import { ETHEREUM_WIDGET_CONFIG } from '../config';
 import { WalletTypes } from '@lidofinance/wallets-testing-wallets';
-import { Page, test, expect } from '@playwright/test';
+import { test, expect } from '@playwright/test';
 import { BrowserService } from '@lidofinance/browser-service';
 import { WidgetPage } from '../pages/widget.page';
 import { waitForTextContent } from '../utils/helpers';
 
 export interface TxConfig {
   txAmount: string;
-  tokenAddress?: string;
-  mappingSlot?: number;
 }
 
 export class WidgetService {
-  page: Page;
   widgetPage: WidgetPage;
 
   constructor(private browserService: BrowserService) {
-    this.page = this.browserService.getBrowserContextPage();
-    this.widgetPage = new WidgetPage(this.page);
+    this.widgetPage = new WidgetPage(
+      this.browserService.getBrowserContextPage(),
+    );
   }
 
   async navigate() {
     await test.step('Navigate to Ethereum widget', async () => {
-      await this.page.goto(ETHEREUM_WIDGET_CONFIG.url);
+      await this.widgetPage.goto(ETHEREUM_WIDGET_CONFIG.url);
     });
   }
 
@@ -30,11 +28,11 @@ export class WidgetService {
     const walletPage = this.browserService.getWalletPage();
 
     await test.step(`Connect wallet ${walletPage.walletConfig.WALLET_NAME}`, async () => {
-      await this.page.waitForTimeout(2000);
+      await this.widgetPage.page.waitForTimeout(2000);
       // If wallet connected -> return
       if ((await this.widgetPage.connectBtn.count()) === 0) return;
       await this.widgetPage.connectBtn.first().click();
-      await this.page.waitForTimeout(2000);
+      await this.widgetPage.page.waitForTimeout(2000);
       // If Stake submit button is displayed -> return
       if ((await this.widgetPage.stakeSubmitBtn.count()) > 0) return;
 
@@ -48,7 +46,7 @@ export class WidgetService {
       switch (walletPage.walletConfig.WALLET_TYPE) {
         case WalletTypes.EOA: {
           const [connectWalletPage] = await Promise.all([
-            this.page.context().waitForEvent('page'),
+            this.widgetPage.waitForPage(),
             walletButton.dblclick(),
           ]);
           await walletPage.connectWallet(connectWalletPage);
@@ -58,7 +56,9 @@ export class WidgetService {
           await walletButton.click();
           await this.widgetPage.copyWcUrlBtn.click();
           await walletPage.connectWallet(
-            await this.page.evaluate(() => navigator.clipboard.readText()),
+            await this.widgetPage.page.evaluate(() =>
+              navigator.clipboard.readText(),
+            ),
           );
           break;
         }
@@ -66,14 +66,12 @@ export class WidgetService {
     });
 
     await test.step('Check the widget after wallet connection', async () => {
-      expect(
-        await this.page.waitForSelector('data-testid=stakeSubmitBtn'),
-      ).not.toBeNaN();
+      await this.widgetPage.stakeSubmitBtn.waitFor({ timeout: 90000 });
       await this.widgetPage.headerAccountSection.click();
       expect(await this.widgetPage.providerName.textContent()).toContain(
         walletPage.walletConfig.CONNECTED_WALLET_NAME,
       );
-      await this.page.locator('div[role="dialog"] button').nth(0).click();
+      await this.widgetPage.closeAccountModal();
     });
   }
 
@@ -81,12 +79,9 @@ export class WidgetService {
     await test.step('Do staking', async () => {
       await waitForTextContent(this.widgetPage.ethAvailableToStakeValue);
       await this.widgetPage.stakeInput.fill(txConfig.txAmount);
-      await this.page.waitForSelector(
-        'button[data-testid="stakeSubmitBtn"]:not([disabled])',
-        { timeout: 15000 },
-      );
+      await this.widgetPage.enabledStakeSubmitBtn.waitFor({ timeout: 15000 });
       const [walletSignPage] = await Promise.all([
-        this.page.context().waitForEvent('page', { timeout: 180000 }),
+        this.widgetPage.waitForPage(180000),
         this.widgetPage.stakeSubmitBtn.click(),
       ]);
